Telagram-Bot-using-GenAI-main/requirements.txt
python-dotenv
aiogram==2.5
cohere